# docs/swagger.yaml
basePath: /
definitions:
...
  api.LogEvent:
    properties:
      caller:
        type: string
      core_id:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      event:
        type: string
      level:
        type: string
      message:
        type: string
      ts:
        format: int64
        type: integer
    type: object
  api.LogEventFilter:
    properties:
      caller:
        type: string
      core_id:
        type: string
      data:
        additionalProperties:
          type: string
        type: object
      event:
        type: string
      level:
        type: string
      message:
        type: string
    type: object
  api.LogEventFilters:
    properties:
      filters:
        items:
          $ref: '#/definitions/api.LogEventFilter'
        type: array
    type: object
...
  api.MediaEvent:
    properties:
      action:
        type: string
      name:
        type: string
      names:
        items:
          type: string
        type: array
      ts:
        type: integer
    type: object
...
  api.ProcessEvent:
    properties:
      core_id:
        type: string
      domain:
        type: string
      line:
        type: string
      pid:
        type: string
      progress:
        $ref: '#/definitions/api.ProcessProgress'
      ts:
        type: integer
      type:
        type: string
    type: object
  api.ProcessEventFilter:
    properties:
      core_id:
        type: string
      domain:
        type: string
      pid:
        type: string
      type:
        type: string
    type: object
  api.ProcessEventFilters:
    properties:
      filters:
        items:
          $ref: '#/definitions/api.ProcessEventFilter'
        type: array
    type: object
...
  api.ProcessProgress:
    properties:
      input:
        items:
          $ref: '#/definitions/api.ProcessProgressInput'
        type: array
      output:
        items:
          $ref: '#/definitions/api.ProcessProgressOutput'
        type: array
      speed:
        type: number
      time:
        type: number
    type: object
  api.ProcessProgressInput:
    properties:
      avstream:
        $ref: '#/definitions/api.ProcessProgressInputAVstream'
      bitrate:
        type: number
      fps:
        type: number
      id:
        type: string
      type:
        type: string
    type: object
  api.ProcessProgressInputAVstream:
    properties:
      drop:
        type: integer
      dup:
        type: integer
      enabled:
        type: boolean
      enc:
        type: integer
      looping:
        type: boolean
      time:
        type: integer
    type: object
  api.ProcessProgressOutput:
    properties:
      bitrate:
        type: number
      fps:
        type: number
      id:
        type: string
      q:
        type: number
      type:
        type: string
    type: object
...
info:
  contact:
    email: hello@datarhei.com
    name: datarhei Core Support
    url: https://www.datarhei.com
  description: Expose REST API for the datarhei Core
  license:
    name: Apache 2.0
    url: https://github.com/datarhei/core/v16/blob/main/LICENSE
  title: datarhei Core API
  version: "3.0"
paths:
...
  /api/v3/cluster/events:
    post:
      consumes:
      - application/json
      description: Stream of log events of whats happening on each node in the cluster
      operationId: cluster-3-events-log
      parameters:
      - description: Event filters
        in: body
        name: filters
        schema:
          $ref: '#/definitions/api.LogEventFilters'
      produces:
      - text/event-stream
      - application/x-json-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LogEvent'
      security:
      - ApiKeyAuth: []
      summary: Stream of log events
      tags:
      - v16.?.?
...
  /api/v3/events:
    post:
      consumes:
      - application/json
      description: Stream of log event of whats happening in the core
      operationId: events-3-log
      parameters:
      - description: Event filters
        in: body
        name: filters
        schema:
          $ref: '#/definitions/api.LogEventFilters'
      produces:
      - text/event-stream
      - application/x-json-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.LogEvent'
      security:
      - ApiKeyAuth: []
      summary: Stream of log events
      tags:
      - v16.?.?
  /api/v3/events/media/{type}:
    post:
      consumes:
      - application/json
      description: Stream of media event of whats happening in the filesystems
      operationId: events-3-media
      parameters:
      - description: glob pattern for media names
        in: query
        name: glob
        type: string
      produces:
      - application/x-json-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.MediaEvent'
      security:
      - ApiKeyAuth: []
      summary: Stream of media events
      tags:
      - v16.?.?
  /api/v3/events/process:
    post:
      consumes:
      - application/json
      description: Stream of process event of whats happening in the processes
      operationId: events-3-process
      parameters:
      - description: Event filters
        in: body
        name: filters
        schema:
          $ref: '#/definitions/api.ProcessEventFilters'
      produces:
      - application/x-json-stream
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.ProcessEvent'
      security:
      - ApiKeyAuth: []
      summary: Stream of process events
      tags:
      - v16.?.?
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ApiRefreshKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  Auth0KeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"